+++ src/test/java/testsuite/Test15.java
@@ -27,62 +27,57 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
  * OF SUCH DAMAGE.
  */
-package purejavacomm.testsuite;
+package testsuite;
 
 import purejavacomm.SerialPortEvent;
 import purejavacomm.SerialPortEventListener;
 
-public class Test13 extends TestBase {
+public class Test15 extends TestBase {
 	static void run() throws Exception {
 
 		try {
-			begin("Test13 - enableReceiveThreshold(0)");
+			int timeout = 100;
+			begin("Test15 - treshold disabled, timeout == " + timeout);
 			openPort();
-
 			m_Out = m_Port.getOutputStream();
 			m_In = m_Port.getInputStream();
 
-			final byte[] txbuffer = new byte[10];
+			final byte[] txbuffer = new byte[1000];
 			final byte[] rxbuffer = new byte[txbuffer.length];
 
 			m_Port.enableReceiveTimeout(100);
-			m_Port.enableReceiveThreshold(0);
-			int totalN = 10;
-			int bytesN = 8;
-			{
-				long totalT = 0;
-				for (int i = 0; i < totalN; i++) {
-					m_Out.write(txbuffer, 0, bytesN);
-					sleep(100); // give the data some time to loop back
-					{ // ask for 10 but expect to get back immediately with the 8 bytes that are available
-						long T0 = System.currentTimeMillis();
-						int n = m_In.read(rxbuffer, 0, 10);
-						long T1 = System.currentTimeMillis();
-						totalT += T1 - T0;
-						if (n != 8)
-							fail("did not get all data back, got only " + n + " bytes");
-					}
-				}
-				if (totalT / totalN > 1)
-					fail("read did not return immediately, it took " + totalT / totalN + " msec on average to read " + bytesN + " bytes");
+			m_Port.disableReceiveThreshold();
 
-			}
 			{
-				long totalT = 0;
-				for (int i = 0; i < totalN; i++) {
-					{ // ask for 10 but expect to get back immediately with the 0 bytes that are available
-						long T0 = System.currentTimeMillis();
-						int n = m_In.read(rxbuffer, 0, 10);
-						long T1 = System.currentTimeMillis();
-						totalT += T1 - T0;
-						if (n != 0)
-							fail("was expecting 0 bytes, but got " + n + " bytes");
-					}
-				}
-				if (totalT / totalN > 1)
-					fail("read did not return immediately, it took " + totalT / totalN + " msec");
+				long T0 = System.currentTimeMillis();
+				int n = m_In.read(rxbuffer, 0, 10);
+				long T1 = System.currentTimeMillis();
+				if (n != 0)
+					fail("was expecting 0 bytes, but got " + n + " bytes");
+				int timeLo = timeout;
+				int timeHi = timeout * 110 / 100;
+				int time = (int) (T1 - T0);
+				if (time < timeLo)
+					fail("timed out early, was expecting  " + timeLo + " but got " + time + " msec");
+				if (time > timeHi)
+					fail("timed out late, was expecting  " + timeHi + " but got " + time + " msec");
 			}
 
+			{
+				m_Out.write(txbuffer, 0, 1000); // at 9600 this should take about 1 sec 
+				sleep(50); // give time to about 50 chars to loop back
+				long T0 = System.currentTimeMillis();
+				int n = m_In.read(rxbuffer, 0, 1000);
+				long T1 = System.currentTimeMillis();
+				int time = (int) (T1 - T0);
+				int etime= n * 150/100; // at 9600
+				if (time > etime)
+					fail("expected read to return in " + etime+" but it took " + time + " msec and returned "+n +" bytes");
+				if (n < 10)
+					fail("was expecting at least 900 bytes, but got " + n + " bytes");
+			}
+
+			
 			finishedOK();
 		} finally {
 			closePort();
