+++ src/test/java/testsuite/Test4.java
@@ -27,41 +27,35 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
  * OF SUCH DAMAGE.
  */
-package purejavacomm.testsuite;
+package testsuite;
 
 import purejavacomm.SerialPortEvent;
 import purejavacomm.SerialPortEventListener;
 
-public class Test6 extends TestBase {
+public class Test4 extends TestBase {
 	private static Exception m_Exception = null;
 	private static Thread m_Receiver;
 	private static Thread m_Transmitter;
 
 	static void run() throws Exception {
 		try {
-			begin("Test6 - threshold + timeout");
+			begin("Test4 - indefinite blocking");
 			openPort();
-			m_Port.setSerialPortParams(230400, 8, 1, 0);
-			//m_In = purejavacomm.RawStream.getInputStream(m_Port);
 			// receiving thread
 			m_Receiver = new Thread(new Runnable() {
 				public void run() {
 					try {
 						sync(2);
-						m_Port.enableReceiveThreshold(4);
-						m_Port.enableReceiveTimeout(10000);
-						//purejavacomm.RawStream.configureThresholdTimeout(m_Port, 4, 2000);
-						byte[] b = new byte[4];
-						for (int i = 0; i < 1000; i++) {
-							long T0 = System.currentTimeMillis();
-							int n = m_In.read(b);
-							long dT = System.currentTimeMillis() - T0;
-							if (n != 4)
-								fail("read did not get 4 bytes as expected, got %d ", n);
-							if (dT >= 1000)
-								fail("read timed out though we got 4 bytes "+dT);
-						}
-
+						long T0 = System.currentTimeMillis();
+						byte[] b = { 0 };
+						int n = m_In.read(b);
+						long dT = System.currentTimeMillis() - T0;
+						if (n != 1)
+							fail("read did not block, read returned %d", n);
+						if (b[0] != 73)
+							fail("read did not get looped back '73' got '%d'", b[0]);
+						if (dT < 10000)
+							fail("read did not block for 10000 msec, received loopback in %d msec", dT);
 					} catch (InterruptedException e) {
 					} catch (Exception e) {
 						if (m_Exception == null)
@@ -77,8 +71,9 @@ public class Test6 extends TestBase {
 				public void run() {
 					try {
 						sync(2);
-						for (int i = 0; i < 1000; i++)
-							m_Out.write(new byte[4]);
+						sleep(10000);
+						m_Out.write(73);
+
 					} catch (InterruptedException e) {
 					} catch (Exception e) {
 						e.printStackTrace();
@@ -90,9 +85,8 @@ public class Test6 extends TestBase {
 				};
 			});
 
-			m_Transmitter.start();
-			sleep(100);
 			m_Receiver.start();
+			m_Transmitter.start();
 
 			while (m_Receiver.isAlive() || m_Transmitter.isAlive()) {
 				sleep(100);
