+++ src/test/java/testsuite/Test5.java
@@ -27,57 +27,72 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
  * OF SUCH DAMAGE.
  */
-package purejavacomm.testsuite;
+package testsuite;
 
 import purejavacomm.SerialPortEvent;
 import purejavacomm.SerialPortEventListener;
 
-public class Test15 extends TestBase {
+public class Test5 extends TestBase {
+	private static Exception m_Exception = null;
+	private static Thread m_Receiver;
+	private static Thread m_Transmitter;
+
 	static void run() throws Exception {
-
 		try {
-			int timeout = 100;
-			begin("Test15 - treshold disabled, timeout == " + timeout);
+			begin("Test5 - timeout");
 			openPort();
-			m_Out = m_Port.getOutputStream();
-			m_In = m_Port.getInputStream();
+			// receiving thread
+			m_Receiver = new Thread(new Runnable() {
+				public void run() {
+					try {
+						sync(2);
+						m_Port.disableReceiveThreshold();
+						m_Port.enableReceiveTimeout(1000);
+						long T0 = System.currentTimeMillis();
+						byte[] b = { 0 };
+						int n = m_In.read(b);
+						long dT = System.currentTimeMillis() - T0;
+						if (n != 0)
+							fail("read did not time out as expected, read returned %d > 0", n);
+						if (dT < 1000 - 10)
+							fail("read timed out early, expected 1000 msec, got %d msec", dT);
+						if (dT > 1100)
+							fail("read timed out with suspicious delay, expected 1000 msec, got %d msec", dT);
+					} catch (InterruptedException e) {
+					} catch (Exception e) {
+						if (m_Exception == null)
+							m_Exception = e;
+						m_Receiver.interrupt();
+						m_Transmitter.interrupt();
+					}
+				};
+			});
 
-			final byte[] txbuffer = new byte[1000];
-			final byte[] rxbuffer = new byte[txbuffer.length];
+			// sending thread
+			m_Transmitter = new Thread(new Runnable() {
+				public void run() {
+					try {
+						sync(2);
+					} catch (InterruptedException e) {
+					} catch (Exception e) {
+						e.printStackTrace();
+						if (m_Exception == null)
+							m_Exception = e;
+						m_Receiver.interrupt();
+						m_Transmitter.interrupt();
+					}
+				};
+			});
 
-			m_Port.enableReceiveTimeout(100);
-			m_Port.disableReceiveThreshold();
+			m_Receiver.start();
+			m_Transmitter.start();
 
-			{
-				long T0 = System.currentTimeMillis();
-				int n = m_In.read(rxbuffer, 0, 10);
-				long T1 = System.currentTimeMillis();
-				if (n != 0)
-					fail("was expecting 0 bytes, but got " + n + " bytes");
-				int timeLo = timeout;
-				int timeHi = timeout * 110 / 100;
-				int time = (int) (T1 - T0);
-				if (time < timeLo)
-					fail("timed out early, was expecting  " + timeLo + " but got " + time + " msec");
-				if (time > timeHi)
-					fail("timed out late, was expecting  " + timeHi + " but got " + time + " msec");
+			while (m_Receiver.isAlive() || m_Transmitter.isAlive()) {
+				sleep(100);
 			}
 
-			{
-				m_Out.write(txbuffer, 0, 1000); // at 9600 this should take about 1 sec 
-				sleep(50); // give time to about 50 chars to loop back
-				long T0 = System.currentTimeMillis();
-				int n = m_In.read(rxbuffer, 0, 1000);
-				long T1 = System.currentTimeMillis();
-				int time = (int) (T1 - T0);
-				int etime= n * 150/100; // at 9600
-				if (time > etime)
-					fail("expected read to return in " + etime+" but it took " + time + " msec and returned "+n +" bytes");
-				if (n < 10)
-					fail("was expecting at least 900 bytes, but got " + n + " bytes");
-			}
-
-			
+			if (m_Exception != null)
+				throw m_Exception;
 			finishedOK();
 		} finally {
 			closePort();
