+++ src/test/java/testsuite/Test6.java
@@ -27,35 +27,40 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
  * OF SUCH DAMAGE.
  */
-package purejavacomm.testsuite;
+package testsuite;
 
 import purejavacomm.SerialPortEvent;
 import purejavacomm.SerialPortEventListener;
 
-public class Test7 extends TestBase {
+public class Test6 extends TestBase {
 	private static Exception m_Exception = null;
 	private static Thread m_Receiver;
 	private static Thread m_Transmitter;
-	private static volatile long m_T0;
 
 	static void run() throws Exception {
 		try {
-			begin("Test7 - threshold");
+			begin("Test6 - threshold + timeout");
 			openPort();
+			m_Port.setSerialPortParams(230400, 8, 1, 0);
+			//m_In = purejavacomm.RawStream.getInputStream(m_Port);
 			// receiving thread
 			m_Receiver = new Thread(new Runnable() {
 				public void run() {
 					try {
 						sync(2);
-						m_Port.enableReceiveThreshold(7);
-						m_Port.disableReceiveTimeout();
-						byte[] b = new byte[8];
-						int n = m_In.read(b);
-						long dT = System.currentTimeMillis() - m_T0;
-						if (n != 7)
-							fail("read did not get 7 bytes as expected, got %d", n);
-						if (dT < 10000)
-							fail("timed out in %d though we got 7 bytes",dT);
+						m_Port.enableReceiveThreshold(4);
+						m_Port.enableReceiveTimeout(10000);
+						//purejavacomm.RawStream.configureThresholdTimeout(m_Port, 4, 2000);
+						byte[] b = new byte[4];
+						for (int i = 0; i < 1000; i++) {
+							long T0 = System.currentTimeMillis();
+							int n = m_In.read(b);
+							long dT = System.currentTimeMillis() - T0;
+							if (n != 4)
+								fail("read did not get 4 bytes as expected, got %d ", n);
+							if (dT >= 1000)
+								fail("read timed out though we got 4 bytes "+dT);
+						}
 
 					} catch (InterruptedException e) {
 					} catch (Exception e) {
@@ -72,10 +77,8 @@ public class Test7 extends TestBase {
 				public void run() {
 					try {
 						sync(2);
-						m_T0 = System.currentTimeMillis();
-						sleep(10000);
-						m_Out.write(new byte[7]);
-
+						for (int i = 0; i < 1000; i++)
+							m_Out.write(new byte[4]);
 					} catch (InterruptedException e) {
 					} catch (Exception e) {
 						e.printStackTrace();
@@ -87,8 +90,9 @@ public class Test7 extends TestBase {
 				};
 			});
 
-			m_Receiver.start();
 			m_Transmitter.start();
+			sleep(100);
+			m_Receiver.start();
 
 			while (m_Receiver.isAlive() || m_Transmitter.isAlive()) {
 				sleep(100);
